fix: fix wins
