+++ src/winners/winners.ts
@@ -4,16 +4,15 @@ import { websocketResponse } from '../models/models';
 
 export const updateWinners = (winnerId: number) => {
   const user = getUser(winnerId);
-  const winnerUpdate = db.winners.map((winner) => {
-    winner.name === user.name ? winner.wins++ : winner.wins
-    return winner;
-  });
-  db.winners = winnerUpdate
-  db.winners.sort((a, b) => a.wins - b.wins);
+  const updatedWins = db.winners.map(({ name, wins }) => (
+    name === user.name ? { name, wins: wins + 1 } : { name, wins }
+  ));
+  db.winners = updatedWins;
+  db.winners.sort((a, b) => a.wins - b.wins).slice(10);
 };
 
 export const winnersUpdate = () => {
   const resp = websocketResponse.Win;
   resp.data = JSON.stringify(db.winners);
   return JSON.stringify(resp);
-}
+};
